v0.8
Added feature map saving (oops) -- feels clumsy implementation. Changed
the string feature keys to numbers, but just now realized a way to
preserve them; alas! Lessons learned.

Also fixed the ending scenes not displaying. combatMenu wasn't being
cleared because I didn't think it was important, but the way I had
love.update set up, it was interrupting the scene displays.

When you heal either in or outside of combat, the proper amount healed
should be displayed, rather than the new total HP.

Cut all the upgrade costs in half. The game isn't long, and isn't
supposed to be, and pretty much the only way to get through in a
reasonable amount of time was to lower those costs. There's a lot of
scrap in the treasure chests anyway and the boss isn't tough. Will do a
start-from-scratch run through later today to make sure it's
completable in a good time frame.