v0.9
Okay the costs are really adjusted this time. No grinding should be
required; there's plenty of scrap in the chests and you'll get some
more as you go along.

Added an intro screen.

Added some sound effects! Not all the effects there should be. There
are more in the folder than I'm currently using.

Pulled the "originals" folder out, because it was large and pointless
to package with this.

I think this is basically done for my purposes for now.